bug fix for Hr and Store

+++ HumanitarianAssistance.Domain/Entities/Store/StoreItemPurchase.cs
@@ -25,7 +25,7 @@ namespace HumanitarianAssistance.Domain.Entities.Store
         [Required]
         public int Currency { get; set; }
         public int UnitType { get; set; }
-        public long UnitCost { get; set; }
+        public double UnitCost { get; set; }
         public int Quantity { get; set; }
 
         public bool ApplyDepreciation { get; set; }
